refactor code, respond to feedback
